refactor: split smalruby_ruby VM files by responsibility
- index.js (327→156 lines): extract block definitions to getBlocks()
  in block-definitions.js, keep only class shell + execution dispatch
- method-executors.js (295→227 lines): remove withBlock executors
- block-method-executors.js (81 lines, new): array/number withBlock
  executors with shared setBlockParam helper
- block-definitions.js: add getBlocks() with all block entries,
  organized by class (String, Array, Hash, Number, blockParam, returnValue)

File structure now clearly separates:
- block-definitions.js: what blocks exist (data + getInfo blocks)
- method-executors.js: COMMAND method execution (per class)
- block-method-executors.js: C-shape block execution (per class)
- index.js: orchestration only
